final commit
Probably and hopefully only minor changes. This is somewhat the working
version at the time of thesis submission

MInor changes were probably made to try to make it work for the git
repository on amazon. But these didn't work out.

The problem is that getting that much data from that bad a repository is
way too difficult to do using JDBC